geometry: reduce size of node by 8 bytes (size_t -> uint32_t)
This commit replaces the use of size_t (the index type of the input
point list) with an explicit uint32_t type on the basis that it is
not currently feasible to consider a point cloud with more than
2**32 points.

uint32_t is used to follow the STL convention of a vector's size_type
being unsigned.

+++ tmc3/PCCTMC3Common.h
@@ -50,8 +50,10 @@ const uint32_t PCCTMC3AdaptiveDataModelAlphabetMaxSize = 2047;
 struct PCCOctree3Node {
   // 3D position of the current node's origin (local x,y,z = 0).
   PCCVector3<uint32_t> pos;
-  size_t start;
-  size_t end;
+
+  // Range of point indexes spanned by node
+  uint32_t start;
+  uint32_t end;
 };
 
 struct PCCNeighborInfo {
